refactor:WinningLotto class
