+++ src/main/java/lotto/LottoResult.java
@@ -1,4 +1,4 @@
-package domain;
+package lotto;
 
 import prize.Prize;
